refactor: 프로젝트 구조 변경
- import 변경
