+++ src/main/java/com/hyune/raider50g/domain/booking/Raider.java
@@ -1,6 +1,6 @@
 package com.hyune.raider50g.domain.booking;
 
-import com.hyune.raider50g.domain.booking.type.ClassType;
+import com.hyune.raider50g.common.type.ClassType;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import javax.persistence.EnumType;
